+++ dart/geobase/CHANGELOG.md
@@ -12,7 +12,8 @@ NOTE: Version 1.1.0 currently under development (1.1.0-dev.0).
   * Support EWKB dimensionality flags for z and m.
   * CoordRefSys to support getting EPSG code as an integer.
   * Support encoding SRID to EWKB binary data.
-  * Utilities to decode a byte order, a WKB flavor and SRID from WKB bytes.
+  * Utilities to decode a byte order, a WKB flavor, a coordinate type and SRID
+    from WKB bytes.
   * A sample for decoding and encoding EWKB binary data.
 * [Add EWKT support](https://github.com/navibyte/geospatial/issues/29)
   * Support decoding also EWKT data (ignoring an optional srid).